refactor(home): move about us section to server component
