+++ apps/frontend/src/components/server/AboutUsServerSection.tsx
@@ -1,0 +1,38 @@
+import { AboutUsSection } from "@repo/ui/components/Composite"
+import { use } from "react"
+import { getAboutUsInfo } from "@/services/cms/about-us-info/AboutUsInfoService"
+
+const mockItems = [
+  {
+    src: "https://images.unsplash.com/photo-1506744038136-46273834b3fb?w=600",
+    alt: "Mountain Lake",
+    width: 600,
+    height: 400,
+    emoji: "😄",
+  },
+  {
+    src: "https://images.unsplash.com/photo-1465101046530-73398c7f28ca?w=600",
+    alt: "Forest Path",
+    width: 600,
+    height: 400,
+    emoji: "😁",
+  },
+  {
+    src: "https://images.unsplash.com/photo-1519125323398-675f0ddb6308?w=600",
+    alt: "Desert Dunes",
+    width: 600,
+    height: 400,
+    emoji: "😆",
+  },
+]
+
+/**
+ * Server-side component to fetch and render the about us section.
+ *
+ * @returns A about us section component.
+ */
+export const AboutUsServerSection = () => {
+  const { data: aboutUsItems } = use(getAboutUsInfo())
+
+  return <AboutUsSection cards={aboutUsItems.items} items={mockItems} />
+}
